+++ src/components/NewsCard.vue
